requirements.txt
Flask==1.1.2
numpy==1.20.3
scikit_learn==0.24.2
scikit_learn==1.2.1
tensorflow-cpu==2.15.0
tensorflow_intel==2.15.0
Werkzeug==3.0.3